fix(database import): Gracefully handle error to get catalog schemas
