+++ superset/commands/database/create.py
@@ -39,9 +39,9 @@ from superset.commands.database.ssh_tunnel.exceptions import (
     SSHTunnelInvalidError,
 )
 from superset.commands.database.test_connection import TestConnectionDatabaseCommand
+from superset.commands.database.utils import add_permissions
 from superset.daos.database import DatabaseDAO
 from superset.databases.ssh_tunnel.models import SSHTunnel
-from superset.db_engine_specs.base import GenericDBException
 from superset.exceptions import OAuth2RedirectError, SupersetErrorsException
 from superset.extensions import event_logger, security_manager
 from superset.models.core import Database
@@ -99,28 +99,7 @@ class CreateDatabaseCommand(BaseCommand):
                 ).run()
 
             # add catalog/schema permissions
-            if database.db_engine_spec.supports_catalog:
-                catalogs = database.get_all_catalog_names(
-                    cache=False,
-                    ssh_tunnel=ssh_tunnel,
-                )
-                for catalog in catalogs:
-                    security_manager.add_permission_view_menu(
-                        "catalog_access",
-                        security_manager.get_catalog_perm(
-                            database.database_name, catalog
-                        ),
-                    )
-            else:
-                # add a dummy catalog for DBs that don't support them
-                catalogs = [None]
-
-            for catalog in catalogs:
-                try:
-                    self.add_schema_permissions(database, catalog, ssh_tunnel)
-                except GenericDBException:  # pylint: disable=broad-except
-                    logger.warning("Error processing catalog '%s'", catalog)
-                    continue
+            add_permissions(database, ssh_tunnel)
         except (
             SSHTunnelInvalidError,
             SSHTunnelCreateFailedError,
